aprove and create

+++ main/views.py
@@ -1,13 +1,8 @@
-from django.shortcuts import render, redirect
-from .forms import *
-from django.contrib.auth.models import User
+from django.shortcuts import render
 from django.contrib.auth import authenticate, login, logout
-from .models import *
-import datetime
+from . import models
+from django.db.models import Model
 from django.http import HttpResponse, JsonResponse
-from django.views.decorators.http import require_POST, require_GET
-from django.utils import six
-from django.utils.dateparse import parse_time
 
 def main(request):
     results = {}
@@ -16,8 +11,8 @@ def main(request):
 
 def goTo(request, page):
     if page == 'profile':
-        if User.is_authenticated:
-            user = Person.objects.get(user=request.user)
+        if request.user.is_authenticated:
+            user = models.Person.objects.get(user=request.user)
             results = user.get_duties()
             if user.user_type == 0:
                 page += '_for_retired'
@@ -28,25 +23,59 @@ def goTo(request, page):
             elif user.user_type == 3:
                 page += '_for_admin'
             else:
-                return HttpResponse("<H4>Error<H4>")
+                return HttpResponse(status=400)
+        else:
+            return main(request)    
     else:
         results = {}
     return render(request, '{}.html'.format(page), results)
 
 def delete(request, object, object_id):
-    obj = globals()[object].objects.get(id=object_id)
-    obj.delete()
-    return HttpResponse("1")
+    model = getModel(object)
+    if model:
+        obj = model.objects.get(id=object_id)
+        obj.delete()
+        return HttpResponse("1")
+    else: 
+        return HttpResponse(status=400)
 
 def logSystem(request):
     if request.method == "POST":
-        print("ENTRO POST")
         username = request.POST.get("personal_id")
         password = request.POST.get("password")
         user = authenticate(request, username=username, password=password)
         if user is not None:
-            print("user es no none")
             login(request, user)
             return main(request)
     logout(request)
     return goTo(request, 'login')
+
+def createModel(request):
+    if request.method == "POST":
+        model = getModel(request.POST['model'])
+        if model:
+            dic ={}
+            for a in request.POST.keys():
+                if hasattr(model, a):
+                    dic[a]=request.POST[a]
+            model.objects.create(**dic)
+        else:
+            return HttpResponse(status=400)
+    return HttpResponse(model)
+
+def aproveRequest(request):
+    if request.method == "POST":
+        request_id = request.POST['model_id']
+        model = getModel(request.POST['model'])
+        if model:
+            request_obj = model.objects.get(id=request_id)
+            request_obj.approved()
+            return HttpResponse("1")
+    return HttpResponse(status=405)
+
+def getModel(model_name):
+    if hasattr(models, model_name):
+        model = getattr(models, model_name)
+        if issubclass(model, Model):
+            return model
+    return 0
